account app

+++ transactions/models.py
@@ -10,7 +10,7 @@ class Trans(models.Model):
     name = models.CharField(max_length=120)
     bank = models.CharField(max_length=50)
     account_no = models.CharField(max_length=10)
-    amount = models.IntegerField()
+    amount =  models.DecimalField(max_digits=8, decimal_places=2)
     date = models.DateTimeField(auto_now_add = True, help_text = "time of transaction")
     transaction_ID = models.CharField(max_length=50, primary_key = True)
     info = models.CharField(max_length = 120)
